Renamed design to common
